requirements.txt
# Core Framework
streamlit>=1.28.0


# Database
supabase>=2.0.0
postgrest>=0.13.0

# Utilities
python-dotenv>=1.0.0
pycountry>=23.12.11

# Additional dependencies that may be needed
pandas>=2.0.0
plotly>=5.17.0
